CarouselDotsList rework
