+++ src/core/Carousel/CarouselDot.tsx
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 import React from 'react';
 import cx from 'classnames';
+import { useMergedRefs } from '../utils';
 
 type CarouselDotProps = {
   /** Is this dot currently active? */
@@ -18,7 +19,7 @@ type CarouselDotProps = {
 
 /**
  * `CarouselDot` is the actual "dot" component used to activate a slide on clicking.
- * It should be used as a child of `CarouselDots`.
+ * It should be used as a child of `CarouselDotsList`.
  */
 export const CarouselDot = React.forwardRef<
   HTMLButtonElement,
@@ -26,12 +27,38 @@ export const CarouselDot = React.forwardRef<
 >((props, ref) => {
   const {
     isActive,
-    isSmaller,
-    isSmall,
-    isInvisible,
+    // isSmaller,
+    // isSmall,
+    // isInvisible,
     className,
     ...rest
   } = props;
+
+  const justMounted = React.useRef(false);
+  const motionOk = React.useRef(
+    window.matchMedia?.('(prefers-reduced-motion: no-preference)').matches,
+  );
+
+  const dotRef = React.useCallback(
+    (el: HTMLButtonElement) => {
+      if (el && isActive) {
+        el.scrollIntoView({
+          behavior:
+            !justMounted.current && motionOk.current ? 'smooth' : 'auto',
+          block: 'nearest',
+          inline: 'center',
+        });
+      }
+    },
+    [isActive],
+  );
+  const refs = useMergedRefs(ref, dotRef);
+
+  React.useEffect(() => {
+    if (justMounted.current) {
+      justMounted.current = false;
+    }
+  });
 
   return (
     <button
@@ -42,15 +69,16 @@ export const CarouselDot = React.forwardRef<
         'iui-carousel-navigation-dot',
         {
           'iui-active': isActive,
-          'iui-first': isSmaller,
-          'iui-second': isSmall,
-          'iui-invisible': isInvisible,
+          // 'iui-first': isSmaller,
+          // 'iui-second': isSmall,
+          // 'iui-invisible': isInvisible,
         },
         className,
       )}
       aria-selected={isActive}
-      ref={ref}
+      ref={refs}
       {...rest}
+      style={{ transition: 'revert' }} // TODO: fix transition in CSS
     />
   );
 });
